test: address CodeRabbit review on PR #3506
Tighten 8 test assertions per CodeRabbit feedback:

- document/v0: use i64::MAX (outside chrono's range) instead of
  u64::MAX (which casts to -1i64, inside range) so the
  DateTime::from_timestamp_millis None branch + unwrap_or_default
  is actually exercised.
- shielded/builder/shielded_transfer: assert the specific
  "fee + transfer_amount overflows u64" message instead of the
  broader "overflow or exceeds total" or-pattern, so the test
  actually verifies checked_add's overflow branch.
- shielded/builder/shielded_withdrawal + unshield: boundary-fee
  tests now accept Ok as valid (exact-boundary values SHOULD pass
  validation) — only assert that any error message doesn't mention
  "exceeds 1000x" / "below minimum required fee".
- prefunded/add: apply the ops the function actually returned via
  apply_batch_low_level_drive_operations and fetch, rather than
  re-seeding through the dispatcher — so the insert branch's
  GroveDB write is what we verify.
- prefunded/empty: assert GroveDB delete op presence/absence
  directly via a local has_delete_op helper rather than checking
  ops.is_empty() (read/cost ops can satisfy non-empty alone).
- verify/state_transition/token_transition: assert the TokenHistory
  contract id appears as an exact path segment for every
  history-enabled transition (Burn/Freeze/Unfreeze/SetPrice) via
  new assert_path_targets_token_history helper. Replaces the old
  flatten+windows check which could match across segment boundaries.
- verify/state_transition/verify_..._with_proof: rename
  verify_shielded_withdrawal_no_nullifiers_returns_invalid_transition
  to verify_shielded_withdrawal_empty_proof_error since the empty
  proof fails before the first_nullifier guard is reachable.

All 185 touched tests still pass.

+++ packages/rs-dpp/src/document/v0/mod.rs
@@ -314,8 +314,11 @@ mod tests {
         // Timestamps that overflow DateTime should use `.unwrap_or_default()`.
         // This ensures the "unwrap_or_default()" branch of Display is hit.
         let mut doc = minimal_doc();
-        // i64::MAX as u64 — from_timestamp_millis will return None
-        doc.created_at = Some(u64::MAX);
+        // u64::MAX casts to -1i64, which IS inside chrono's range (1 ms before
+        // epoch). Use i64::MAX instead — it exceeds chrono's supported ms
+        // range (~262,000 years) so `from_timestamp_millis` returns None and
+        // the `.unwrap_or_default()` branch is actually exercised.
+        doc.created_at = Some(i64::MAX as u64);
         let s = format!("{}", doc);
         // Must not panic and must contain the created_at prefix
         assert!(s.contains("created_at:"));
